requirements.txt
# Natural Language Processing
transformers==4.28.1
torch==1.13.1
nltk==3.8.1

# Machine Learning
scikit-learn==1.2.2
gensim==4.3.1

# Data Processing
numpy==1.24.3
pandas==2.0.1

# API Interaction (for ChatGPT)
openai==0.27.6

# Text Processing
spacy==3.5.2

# Visualization (for benchmarking results)
matplotlib==3.7.1
seaborn==0.12.2

# Testing
pytest==7.3.1

# Utility
tqdm==4.65.0

# Development Tools
black==23.3.0
flake8==6.0.0
PyYAML~=6.0.2
scipy~=1.10.1